1,温度精确到浮点数。2，增加查询温度接口

+++ plug2.0/app/include/user_temp.h
@@ -10,7 +10,7 @@
 
 
 
-int TEMP_GetTemperature( VOID );
+float TEMP_GetTemperature( VOID );
 VOID TEMP_TempCallBack( VOID );
 
 #endif /* __USER_TEMP_H__ */
